// smart-habit-tracker/src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Cadence = {
  daily: 'daily',
  weekly: 'weekly'
} as const

export type Cadence = (typeof Cadence)[keyof typeof Cadence]